Introduce an interface.
