+++ labelshift/interfaces/__init__.py
@@ -1,0 +1,5 @@
+"""The subpackage with common interfaces.
+
+Note:
+    This file should be empty. Refer to particular submodules.
+"""
